LayoutTests/imported/w3c/web-platform-tests/webstorage/storage_string_conversion.window-expected.txt


FAIL localStorage only stores strings assert_equals: expected "function(){}" but got "function (){}"
FAIL sessionStorage only stores strings assert_equals: expected "function(){}" but got "function (){}"
